+++ aggsender/types/types_test.go
@@ -3,7 +3,6 @@ package types
 import (
 	"fmt"
 	"math/big"
-	"reflect"
 	"testing"
 	"time"
 
@@ -80,40 +79,4 @@ func TestCertificate_String(t *testing.T) {
 
 		require.Equal(t, expected, cert.String())
 	})
-}
-
-func TestCertificateFieldsMatchCertificateInfo(t *testing.T) {
-	// Check that all fields in CertificateInfo are present in Certificate
-	certificateInfoType := reflect.TypeOf(CertificateInfo{})
-
-	for i := range certificateInfoType.NumField() {
-		field := certificateInfoType.Field(i)
-		_, found := reflect.TypeOf(CertificateHeader{}).FieldByName(field.Name)
-		if !found {
-			_, found = reflect.TypeOf(Certificate{}).FieldByName(field.Name)
-		}
-		require.True(t, found, "Field %s is missing in Certificate", field.Name)
-	}
-
-	// Check that all fields in Certificate are present in CertificateInfo
-	certificateHeaderType := reflect.TypeOf(CertificateHeader{})
-
-	// Check that all fields in CertificateHeader are present in CertificateInfo
-	for i := range certificateHeaderType.NumField() {
-		field := certificateHeaderType.Field(i)
-		_, found := certificateInfoType.FieldByName(field.Name)
-		require.True(t, found, "Field %s from CertificateHeader is missing in CertificateInfo", field.Name)
-	}
-
-	// Check that all fields in Certificate are present in CertificateInfo
-	certificateType := reflect.TypeOf(Certificate{})
-
-	for i := range certificateType.NumField() {
-		field := certificateType.Field(i)
-		if field.Name == "Header" {
-			continue
-		}
-		_, found := certificateInfoType.FieldByName(field.Name)
-		require.True(t, found, "Field %s from Certificate is missing in CertificateInfo", field.Name)
-	}
 }
